pledge-india/requirements/local.txt
astroid==1.6.5
certifi==2018.4.16
chardet==3.0.4
defusedxml==0.5.0
Django==2.0.6
django-crispy-forms==1.7.2
django-environ==0.4.4
flake8==3.5.0
idna==2.7
isort==4.3.4
lazy-object-proxy==1.3.1
mccabe==0.6.1
oauthlib==2.1.0
psycopg2==2.7.4
pycodestyle==2.3.1
pyflakes==1.6.0
PyJWT==1.6.4
pylint==1.9.2
python3-openid==3.1.0
pytz==2018.4
requests==2.19.1
requests-oauthlib==1.0.0
six==1.11.0
social-auth-app-django==2.1.0
social-auth-core==1.7.0
urllib3==1.23
wrapt==1.10.11
